Fix UI 3

See merge request hekateai/coke-stores!24

+++ src/modules/shops/cart/CartItem.js
@@ -5,8 +5,11 @@ import { useHistory } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import CartService from '../../../_services/cart';
 import { addCart } from './../../../redux/actions/index';
+import Swal from "sweetalert2"
+import withReactContent from 'sweetalert2-react-content';
 
 const CartItem = ({item,index}) => {
+  const MySwal = withReactContent(Swal)
   const history = useHistory();
   const dispatch = useDispatch();
 
@@ -23,8 +26,19 @@ const CartItem = ({item,index}) => {
   }
 
   const removeCartItem = () => {
-    CartService.remove(index);
-    dispatch(addCart())
+    MySwal.fire({
+      text: "Bạn có đồng ý xóa sản phẩm này khỏi giỏ hàng ?!",
+      icon: 'info',
+      confirmButtonText: 'Đồng ý',
+      showCancelButton: true,
+      cancelButtonText: "Huỷ bỏ"
+    }).then((result)=>{
+        if(result.isConfirmed){
+          CartService.remove(index);
+          dispatch(addCart())
+        }
+    })
+
   }
 
   return (
